+++ Scripts/Mdbc.ps1
@@ -21,12 +21,10 @@
 		nmu - New-MdbcUpdate
 
 	Global variables:
-		$mserver   - the server
-		$mdatabase - the database
-		$m<name>   - collection <name> (for each collection)
-
-	Global variables for collections are especially useful with the tab
-	expansion tool. Conflicts with system variables are highly unlikely.
+		$mserver    - the server
+		$mdatabase  - the database
+		$m<name>    - collection <name> (for each collection)
+		$<operator> - operator shortcuts for JSON-like expressions
 
 	With a large number of collections their names are not displayed. Command
 	Get-Variable m*..* is useful for finding a collection by its name pattern.
@@ -35,23 +33,25 @@
 		# connect to the default server and the database 'test'
 		mdbc
 
-		# get documents from 'files' where Length > 1GB
+		# get from 'files' where Length > 1GB
 		gmd $mfiles (nmq Length -gt 1gb)
+		gmd $mfiles @{Length = @{$gt = 1gb}}
 
 .Parameter ConnectionString
 		Connection string (see the C# driver manual for details).
 		The default is "." which stands for "mongodb://localhost".
 
-.Parameter Database
-		The database name.
-		The default is 'test'.
+.Parameter DatabaseName
+		Database name or wildcard pattern. If it is not resolved to an existing
+		database name then the script prints all database names and exits. The
+		default name is 'test'.
 #>
 
 param
 (
 	[Parameter()]
 	$ConnectionString = '.',
-	$Database = 'test'
+	$DatabaseName = 'test'
 )
 
 Import-Module Mdbc
@@ -65,18 +65,58 @@ Set-Alias -Scope global -Name nmd -Value New-MdbcData
 Set-Alias -Scope global -Name nmq -Value New-MdbcQuery
 Set-Alias -Scope global -Name nmu -Value New-MdbcUpdate
 
+# Query operators
+$global:all = '$all'
+$global:and = '$and'
+$global:elemMatch = '$elemMatch'
+$global:exists = '$exists'
+$global:gt = '$gt'
+$global:gte = '$gte'
+$global:in = '$in'
+$global:lt = '$lt'
+$global:lte = '$lte'
+$global:mod = '$mod'
+$global:ne = '$ne'
+$global:nin = '$nin'
+$global:nor = '$nor'
+$global:not = '$not'
+$global:options = '$options'
+$global:or = '$or'
+$global:regex = '$regex'
+$global:size = '$size'
+$global:type = '$type'
+
+# Update operators
+$global:addToSet = '$addToSet'
+$global:bit = '$bit'
+$global:each = '$each'
+$global:inc = '$inc'
+$global:pop = '$pop'
+$global:pull = '$pull'
+$global:pullAll = '$pullAll'
+$global:push = '$push'
+$global:pushAll = '$pushAll'
+$global:rename = '$rename'
+$global:set = '$set'
+$global:unset = '$unset'
+
 # Server variable
 $global:mserver = Connect-Mdbc $ConnectionString
 Write-Host "Server `$mserver $($mserver.Settings.Server)"
 
 # Database variable
-$global:mdatabase = $mServer.GetDatabase($Database)
-Write-Host "Database `$mdatabase $Database"
+$name = @($mserver.GetDatabaseNames() -like $DatabaseName)
+if ($name.Count -ne 1) {
+	Write-Host "Server databases: $($mserver.GetDatabaseNames())"
+	return
+}
+Write-Host "Database `$mdatabase $name"
+$global:mdatabase = $mserver.GetDatabase($name)
 
 # Collection variables
-$collections = @($mDatabase.GetCollectionNames())
-$global:MaximumVariableCount += $collections.Count
+$collections = @($mdatabase.GetCollectionNames())
 Write-Host "$($collections.Count) collections"
+$global:MaximumVariableCount = 32kb
 foreach($name in $collections) {
 	if (!$name.StartsWith('system.')) {
 		if ($collections.Count -lt 50) { Write-Host "Collection `$m$name" }